+++ CHANGELOG.md
@@ -10,6 +10,8 @@
     - By default, recursively read Markdown files in notes directory. This behaviour can be disabled via the neuronignore plugin.
 - Reduce error verbosity (CLI & web) in the scenario of there being innumerous broken wiki-links.
 - `neuron rib` is renamed to `neuron gen` (neuron has migrated from rib/shake to reflex)
+- Performance improvements:
+  - Improve incremental generation performance (`-w) (#522)
 
 ## Unreleased (v1 + v2)
